error responses for course module

+++ src/modules/courses/services/course.service/destroy.service.ts
@@ -1,3 +1,4 @@
+import { ApiError } from "@point-hub/express-error-handler";
 import { CourseRepository } from "../../repositories/course.repository.js";
 import DatabaseConnection from "@src/database/connection.js";
 
@@ -8,6 +9,10 @@ export class DestroyCourseService {
   }
   public async handle(id: string) {
     const courseRepository = new CourseRepository(this.db);
+
+    const course = await courseRepository.read(id);
+    if (!course) throw new ApiError(404)
+    
     await courseRepository.delete(id);
     return {};
   }
