+++ src/modules/users/services/destroy.service.ts
@@ -1,3 +1,5 @@
+import { ApiError } from "@point-hub/express-error-handler";
+import { UserInterface } from "../entities/user.entity.js";
 import { UserRepository } from "../repositories/user.repository.js";
 import DatabaseConnection from "@src/database/connection.js";
 
@@ -8,6 +10,10 @@ export class DestroyUserService {
   }
   public async handle(id: string) {
     const userRepository = new UserRepository(this.db);
+
+    const user: UserInterface = await userRepository.read(id);
+    if(!user) throw new ApiError(404, { msg: 'user not found' });
+    
     await userRepository.delete(id);
     return {};
   }
